🧪 : 새로 적용된 룰 테스트
